Support for u command

+++ local.sh
@@ -8,17 +8,22 @@ fi
 echo "Running command $COMMAND" 
 
 function check {
-    ./csync2 -p 30860 -K csync2_local.cfg -N local -${COMMAND}r$DEBUG test/local >> csync_local.log 2>&1
+    ./csync2 -p 30860 -K csync2_local.cfg -N local -${COMMAND}r$DEBUG test/local # >> csync_local.log 2>&1
     echo "select * from file; select * from dirty;" | mysql -t -u csync2_local -pcsync2_local csync2_local >> mysql_local.log
     ${PAUSE}
 }
 
-echo "delete from dirty ; delete from file" | mysql -u csync2_local -pcsync2_local csync2_local
-echo "delete from dirty ; delete from file" | mysql -u csync2_peer -pcsync2_peer csync2_peer
-rm csync_local.log mysql_local.log
-rm -rf test/local/*
-rm -rf test/peer/*
-check
+if [ "$COMAMND" != "u" ] ; then 
+    echo "delete from dirty ; delete from file" | mysql -u csync2_local -pcsync2_local csync2_local
+    echo "delete from dirty ; delete from file" | mysql -u csync2_peer -pcsync2_peer csync2_peer
+    rm csync_local.log mysql_local.log
+    rm -rf test/local/*
+    rm -rf test/peer/*
+else 
+    check
+    exit 0;
+fi
+
 echo first > test/local/first.txt
 check
 mkdir test/local/dir1
